tibber_mqtt/requirements.txt
pyTibber>=0.31.0
aiohttp
paho-mqtt
